clean up in Texture for old OpenGL code; defty avoid sucky gl headers with func ptr prefix

+++ branches/freyja_0.9.5/freyja/src/FreyjaRender.h
@@ -327,7 +327,7 @@ public:
 	 ------------------------------------------------------*/
 
 	void SetViewportWindow(long idx, freyja_plane_t win)
-	{ if (idx < mViewportsCount) mViewports[idx].plane = win; }
+	{ if (idx < (long)mViewportsCount) mViewports[idx].plane = win; }
 	/*------------------------------------------------------
 	 * Pre  : 
 	 * Post :
